Major refactor
- Split into data class and utility class
- Moved static functions to project wide utility file
- Enforced SRP on loading
- Introduced logging
